+++ src/OpenTelemetry.AutoInstrumentation.Native/integration_loader.h
@@ -5,6 +5,7 @@
 #include <locale>
 #include <nlohmann/json.hpp>
 #include <string>
+#include <unordered_set>
 #include <vector>
 
 #include "integration.h"
@@ -20,12 +21,18 @@ public:
   LoadIntegrationConfiguration(bool traces_enabled,
                                std::vector<WSTRING> enabled_trace_integration_names,
                                std::vector<WSTRING> disabled_trace_integration_names,
+                               bool metrics_enabled,
+                               std::vector<WSTRING> enabled_metric_integration_names,
+                               std::vector<WSTRING> disabled_metric_integration_names,
                                bool logs_enabled,
                                std::vector<WSTRING> enabled_log_integration_names,
                                std::vector<WSTRING> disabled_log_integration_names)
     : traces_enabled(traces_enabled),
       enabledTraceIntegrationNames(std::move(enabled_trace_integration_names)),
       disabledTraceIntegrationNames(std::move(disabled_trace_integration_names)),
+      metrics_enabled(metrics_enabled),
+      enabledMetricIntegrationNames(std::move(enabled_metric_integration_names)),
+      disabledMetricIntegrationNames(std::move(disabled_metric_integration_names)),
       logs_enabled(logs_enabled),
       enabledLogIntegrationNames(std::move(enabled_log_integration_names)),
       disabledLogIntegrationNames(std::move(disabled_log_integration_names)) {
@@ -34,6 +41,9 @@ public:
   const bool traces_enabled;
   const std::vector<WSTRING> enabledTraceIntegrationNames;
   const std::vector<WSTRING> disabledTraceIntegrationNames;
+  const bool metrics_enabled;
+  const std::vector<WSTRING> enabledMetricIntegrationNames;
+  const std::vector<WSTRING> disabledMetricIntegrationNames;
   const bool logs_enabled;
   const std::vector<WSTRING> enabledLogIntegrationNames;
   const std::vector<WSTRING> disabledLogIntegrationNames;
@@ -60,11 +70,11 @@ void LoadIntegrationsFromStream(
 namespace
 {
     void IntegrationFromJson(const json::value_type& src,
-                         std::vector<IntegrationMethod>& integrationMethods,
+                         std::unordered_set<IntegrationMethod>& integrationMethods,
                          const LoadIntegrationConfiguration& configuration);
 
     void MethodReplacementFromJson(const json::value_type& src, const WSTRING& integrationName,
-                                   std::vector<IntegrationMethod>& integrationMethods);
+                                   std::unordered_set<IntegrationMethod>& integrationMethods);
 
     MethodReference MethodReferenceFromJson(const json::value_type& src, const bool is_target_method,
                                             const bool is_wrapper_method);