+++ test/test-applications/integrations/TestApplication.NServiceBus/TestMessage.cs
@@ -1,4 +1,4 @@
-// <copyright file="InstrumentationType.cs" company="OpenTelemetry Authors">
+// <copyright file="TestMessage.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,11 +14,10 @@
 // limitations under the License.
 // </copyright>
 
-namespace OpenTelemetry.AutoInstrumentation.Instrumentations;
+using NServiceBus;
 
-internal enum InstrumentationType
+namespace TestApplication.NServiceBus;
+
+public class TestMessage : IMessage
 {
-    Trace,
-    Metric,
-    Log
 }